jorney'

+++ str_utils.c
@@ -3,7 +3,7 @@
  * @str: parameter
  * Return: length of string
  */
-ssize_t _strlen(const char *str)
+int _strlen(const char *str)
 {
 	int len = 0, count = 0;
